Theres a lot in here.
I thought about adding a list of all the changes in the commit message
but...well you know.

These are mostly changes to make it more comfortable to edit Clojure,
and really lisp in general, dealing with the repl, and what not. Also a
change from airline to lightline, I wasn't really using airline as much
as I could, and I decided to switch to something smaller/simplier.